refactor: :lipstick: Change the assets icons
